+++ src/features/TextToImage/components/TextToImageOutputAlert.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import * as TextToImage from '@/features/TextToImage';
+import * as TextToImage from '@/features/TextToImage/exports';
 
 import TextToImageServerConnectionAlert from './TextToImageServerConnectionAlert.vue';
 import TextToImageServerSpecificationAlert from './TextToImageServerSpecificationAlert.vue';
